separating the run episode function
